code clean + check if actually placed + more stuff
