+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 token.json
 *.png
 *.jpg
+.vscode
