+++ config.json
@@ -1,81 +1,3 @@
 {
-    "colors": [
-        [255, 255, 255],
-        [196, 196, 196],
-        [136, 136, 136],
-        [85, 85, 85],
-        [34, 34, 34],
-        [0, 0, 0],
-        [0, 102, 0],
-        [34, 177, 76],
-        [2, 190, 1],
-        [148, 224, 68],
-        [251, 255, 91],
-        [229, 217, 0],
-        [230, 190, 12],
-        [229, 149, 0],
-        [160, 106, 66],
-        [153, 83, 13],
-        [99, 60, 31],
-        [107, 0, 0],
-        [159, 0, 0],
-        [229, 0, 0],
-        [187, 79, 0],
-        [255, 117, 95],
-        [255, 196, 159],
-        [255, 223, 204],
-        [255, 167, 209],
-        [207, 110, 228],
-        [236, 8, 236],
-        [130, 0, 128],
-        [2, 7, 99],
-        [0, 0, 234],
-        [4, 75, 255],
-        [101, 131, 207],
-        [54, 186, 255],
-        [0, 131, 199],
-        [0, 211, 221]
-    ],
-    "cdict": [
-        [255, 255, 255],
-        [196, 196, 196],
-        [136, 136, 136],
-        [85, 85, 85],
-        [34, 34, 34],
-        [0, 0, 0],
-        [0, 102, 0],
-        [34, 177, 76],
-        [2, 190, 1],
-        [81, 255, 25],
-        [148, 224, 68],
-        [251, 255, 91],
-        [229, 217, 0],
-        [230, 190, 12],
-        [229, 149, 0],
-        [160, 106, 66],
-        [153, 83, 13],
-        [99, 60, 31],
-        [107, 0, 0],
-        [159, 0, 0],
-        [229, 0, 0],
-        [255, 57, 4],
-        [187, 79, 0],
-        [255, 117, 95],
-        [255, 196, 159],
-        [255, 223, 204],
-        [255, 167, 209],
-        [207, 110, 228],
-        [236, 8, 236],
-        [130, 0, 128],
-        [81, 0, 255],
-        [2, 7, 99],
-        [0, 0, 234],
-        [4, 75, 255],
-        [101, 131, 207],
-        [54, 186, 255],
-        [0, 131, 199],
-        [0, 211, 221],
-        [69, 255, 2]
-    ],
-    "boardId": 13
+    "boardId": 11
 }
