+++ package.json
@@ -1,10 +1,7 @@
 {
   "dependencies": {
-    "canvas": "^2.9.1",
     "jimp": "^0.16.1",
-    "node-dithering": "^0.0.1",
     "puppeteer": "^13.6.0",
-    "rgbquant": "^1.1.2",
     "user-agents": "^1.0.994",
     "websocket": "^1.0.34"
   }
